+++ NEWS.md
@@ -6,7 +6,15 @@
 3. Added *libsodium* as dependency in *README.md* and *dependency-install.sh*.
 4. Fixed *dependency-install.sh*.
 5. Added **ChaCha20 encryption** to *rpsecore-broadcast-h* and *rpsecore-broadcast-c*.
-6. Removed redundant DLL deletion at the end of the broadcast receiver.
+6. Removed redundant DLL deletion at the end of *rpsecore_broadcast_broadcasterLoop()*.
+7. Added found players at the end of *rpse_broadcast_receiverLoop()*
+8. Added *user type* macros in *rpsecore-broadcast.h* and added any other externally-used ones there, too.
+9. Removed redundant code from *rpse_io_str()* and added *first_time_called* for safety.
+10. Added thank you message at the end of the program.
+11. Removed redundant reusable part in the receiver function.
+12. Added mutexes to broadcast loop termination handler.
+13. Replaced broadcast and recevier *do-while* loops with *while* loops.
+14. Made broadcast interval be every 20 seconds for the sake of the network that'll be used.
 
 # Older NEWS information #
 Older news information can be found in the ```ONEWS.md``` archive file.